fix bugs ch1-4

+++ Debug/圆的面积与周长/7.c
@@ -3,11 +3,11 @@
 #define PI 3.14159
 
 int main() {
-    float r,c,s;
-    scanf("%f",&r);
+    double r,c,s;
+    scanf("%lf",&r);
     c=2 * PI * r;
     s=PI * r * r;
-    printf("%.3f\n",c);
-    printf("%.3f",s);
+    printf("%.3lf\n",c);
+    printf("%.3lf",s);
     return 0;
 }
