+++ Debug/圆的面积与周长/9.c
@@ -2,13 +2,13 @@
 #define PI 3.14159
 
 int main() {
-    float r;
-    float c, s;
-    scanf("%f", &r);
+    double r;
+    double c, s;
+    scanf("%lf", &r);
     c = 2 * PI * r;
     s = PI * r * r;
-    printf("%.3f\n", c);
-    printf("%.3f\n", s);
+    printf("%.3lf\n", c);
+    printf("%.3lf\n", s);
     return 0;
 }
